Fixed copying test files with relative paths. JIRA:CXX-2812
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@52013 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ scripts/common/check/check_make_unix.sh
@@ -731,7 +731,7 @@ for x_row in $x_tests; do
             x_copy="$x_src_dir/$i"
             if test -f "$x_copy"  -o  -d "$x_copy"; then
                cp -prf "$x_copy" "$x_work_dir"
-               find "$x_work_dir/$i" -name .svn -print | xargs rm -rf
+               test -d "$x_work_dir/$i" &&  find "$x_work_dir/$i" -name .svn -print | xargs rm -rf
             else
                echo "Warning:  The copied object \"$x_copy\" should be a file or directory!"
                continue 1
